backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2
ta==0.10.2  # Alternative to TA-Lib
python-jose==3.3.0  # For JWT
passlib==1.7.4  # For password hashing
python-multipart==0.0.6  # For form data
asyncio==3.4.3
